+++ TPie/Config/RingSettingsWindow.cs
@@ -254,7 +254,7 @@ namespace TPie.Config
         {
             if (Ring == null) return;
 
-            ImGui.SetNextWindowSize(new(94 * _scale, 120 * _scale));
+            ImGui.SetNextWindowSize(new(94 * _scale, 140 * _scale));
 
             if (ImGui.BeginPopup("##TPie_Add_Item_Menu"))
             {
@@ -283,6 +283,11 @@ namespace TPie.Config
                 if (ImGui.Selectable("Game Macro"))
                 {
                     elementToAdd = new GameMacroElement();
+                }
+
+                if (ImGui.Selectable("Nested Ring"))
+                {
+                    elementToAdd = new NestedRingElement();
                 }
 
                 if (elementToAdd != null)